.ci/bot/update_special_files/requirements.txt
python-gitlab
BeautifulSoup4
Jinja2
click
mistune